Adicionado a funcionalidade para listar despesas no dashboard.

+++ Persistencia/DAL/Tabelas/DashboardDAL.cs
@@ -6,7 +6,7 @@ namespace Persistencia.DAL.Tabelas
     public class DashboardDAL
     {
         EFContext context = new EFContext();
-        public double ObterSomaDespesasMensal(int mes, int ano)
+        public decimal ObterSomaDespesasMensal(int mes, int ano)
         {
             var somaDespesasFixas = context.DespesasFixas.Where(x => x.DataHora.Month == mes && x.DataHora.Year == ano).Sum(x => x.Valor);
             var somaDespesasParceladas = context.DespesasParceladas.Where(x => x.DataHora.Month == mes && x.DataHora.Year == ano).Sum(x => x.Valor);
