+++ Persistencia/Migrations/EFContextModelSnapshot.cs
@@ -31,7 +31,7 @@ namespace Persistencia.Migrations
 
                     b.Property<string>("NomeItem");
 
-                    b.Property<double>("Valor");
+                    b.Property<decimal>("Valor");
 
                     b.HasKey("DespesaFixaId");
 
@@ -48,7 +48,7 @@ namespace Persistencia.Migrations
 
                     b.Property<string>("NomeItem");
 
-                    b.Property<double>("Valor");
+                    b.Property<decimal>("Valor");
 
                     b.HasKey("DespesaMensalId");
 
@@ -71,7 +71,7 @@ namespace Persistencia.Migrations
 
                     b.Property<bool>("StatusPagamento");
 
-                    b.Property<double>("Valor");
+                    b.Property<decimal>("Valor");
 
                     b.Property<double>("ValorParcela");
