+++ Servico/Tabelas/DashboardServico.cs
@@ -5,7 +5,7 @@ namespace Servico.Tabelas
     public class DashboardServico
     {
         DashboardDAL dashboardDAL = new DashboardDAL();
-        public double ObterSomaDespesasMensal(int mes, int ano)
+        public decimal ObterSomaDespesasMensal(int mes, int ano)
         {
             return dashboardDAL.ObterSomaDespesasMensal(mes, ano);
         }
